database:  upgrade user ID from Integer to BigInteger to support 64 bit user identifiers

+++ Natsuki/modules/sql/afk_sql.py
@@ -1,6 +1,6 @@
 import threading
 
-from sqlalchemy import Boolean, Column, Integer, UnicodeText
+from sqlalchemy import Boolean, Column, BigInteger, UnicodeText
 
 from Natsuki.modules.sql import BASE, SESSION
 
@@ -8,7 +8,7 @@ from Natsuki.modules.sql import BASE, SESSION
 class AFK(BASE):
     __tablename__ = "afk_users"
 
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     is_afk = Column(Boolean)
     reason = Column(UnicodeText)
