+++ Natsuki/modules/sql/warns_sql.py
@@ -1,6 +1,6 @@
 import threading
 
-from sqlalchemy import Boolean, Column, Integer, String, UnicodeText, distinct, func
+from sqlalchemy import Boolean, Column, Integer, String, UnicodeText, distinct, func, BigInteger
 from sqlalchemy.dialects import postgresql
 
 from Natsuki.modules.sql import BASE, SESSION
@@ -9,7 +9,7 @@ from Natsuki.modules.sql import BASE, SESSION
 class Warns(BASE):
     __tablename__ = "warns"
 
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     chat_id = Column(String(14), primary_key=True)
     num_warns = Column(Integer, default=0)
     reasons = Column(postgresql.ARRAY(UnicodeText))
